[AIRFLOW-3471] Move XCom out of models.py
